double заменены на BigDecimal

+++ src/EmplPerson.java
@@ -1,13 +1,15 @@
+import java.math.BigDecimal;
+
 /**
  * Хранение данных о сотрудника
  */
 class EmplPerson {
 
     private String lastName; //Фамилия
-    private double salary; //зарплата
+    private BigDecimal salary; //зарплата
     private int id; //Идентификатор
 
-    EmplPerson(int id, String lastName, double salary) {
+    EmplPerson(int id, String lastName, BigDecimal salary) {
         this.id = id;
         this.lastName = lastName;
         this.salary = salary;
@@ -18,7 +20,7 @@ class EmplPerson {
         return lastName + ";"  + salary;
     }
 
-    public double getSalary() {return salary;}
+    public BigDecimal getSalary() {return salary;}
     public String getLastName() {return lastName;}
     public int getId() {return id;}
